requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Mail==0.10.0
groq==1.0.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.2
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
scikit-learn==1.8.0
scipy==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.5
